fix various UI issues
