+++ opensaas-sh/app_diff/src/landing-page/contentSections.tsx.diff
@@ -115,7 +115,7 @@
 +  {
 +    name: 'Jonathan Cocharan',
 +    role: 'Entrepreneur',
-+    avatarSrc: 'https://pbs.twimg.com/profile_images/1910056203863883776/jtfVWaEG_400x400.jpg',
++    avatarSrc: 'https://pbs.twimg.com/profile_images/1950172296376639488/sZ0JIqfR_400x400.jpg',
 +    socialUrl: 'https://twitter.com/JonathanCochran',
 +    quote:
 +      'In just 6 nights... my SaaS app is live 🎉! Huge thanks to the amazing @wasplang community 🙌 for their guidance along the way. These tools are incredibly efficient 🤯!',
