+++ template/app/src/client/Main.css
@@ -158,7 +158,7 @@
     --muted-foreground: 0 0% 63.9%;
     --accent: 31 57% 93%;
     --accent-foreground: 210 100% 13%;
-    --destructive: 0 62.8% 30.6%;
+    --destructive: 0 62.8% 65%;
     --destructive-foreground: 0 0% 98%;
     --success: 141 71% 48%;
     --success-foreground: 0 0% 98%;
