+++ template/app/src/user/DropdownUser.tsx
@@ -1,11 +1,18 @@
 import { ChevronDown, User } from 'lucide-react';
+import { useState } from 'react';
 import { type User as UserEntity } from 'wasp/entities';
 import { DropdownMenu, DropdownMenuContent, DropdownMenuTrigger } from '../components/ui/dropdown-menu';
 import { UserMenuItems } from './UserMenuItems';
 
 const DropdownUser = ({ user }: { user: Partial<UserEntity> }) => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const handleItemClick = () => {
+    setIsOpen(false);
+  };
+
   return (
-    <DropdownMenu>
+    <DropdownMenu open={isOpen} onOpenChange={setIsOpen}>
       <DropdownMenuTrigger asChild>
         <button className='flex items-center gap-4 duration-300 ease-in-out text-foreground hover:text-primary transition-colors'>
           <span className='hidden text-right lg:block'>
@@ -16,7 +23,7 @@ const DropdownUser = ({ user }: { user: Partial<UserEntity> }) => {
         </button>
       </DropdownMenuTrigger>
       <DropdownMenuContent className='w-62.5'>
-        <UserMenuItems user={user} />
+        <UserMenuItems user={user} onItemClick={handleItemClick} />
       </DropdownMenuContent>
     </DropdownMenu>
   );
