+++ template/app/src/user/UserMenuItems/index.ts
@@ -1,0 +1,2 @@
+export { MobileMenuItems } from './MobileMenuItems';
+export { UserMenuItems } from './UserMenuItems';
